+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/connecting/SSOLogin.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -15,6 +14,8 @@ import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedProgressIndicator
 import de.connect2x.trixnity.messenger.viewmodel.connecting.AddMatrixAccountState
 import de.connect2x.trixnity.messenger.viewmodel.connecting.SSOLoginViewModel
 
@@ -39,10 +40,15 @@ class SSOLoginViewImpl : SSOLoginView {
             Text(i18n.externalLogin(ssoLoginViewModel.providerName ?: "SSO"))
             Spacer(Modifier.height(20.dp))
         }
+
         Box(Modifier.defaultMinSize(minHeight = 20.dp)) {
             when (state) {
                 AddMatrixAccountState.None -> {}
-                AddMatrixAccountState.Connecting -> LinearProgressIndicator(Modifier.fillMaxWidth())
+                AddMatrixAccountState.Connecting -> ThemedProgressIndicator(
+                    Modifier.fillMaxWidth(),
+                    MaterialTheme.components.linearProgressIndicator
+                )
+
                 is AddMatrixAccountState.Failure ->
                     Text(state.message, color = MaterialTheme.colorScheme.error)