+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/settings/RoomSettingsMemberList.kt
@@ -133,8 +133,10 @@ fun MemberList(
                     )
                 }
             }
-            if (showLoadingSpinner) item(key = "loadingSpinner") {
-                LoadingSpinner()
+            if (showLoadingSpinner) {
+                item(key = "loadingSpinner") {
+                    LoadingSpinner()
+                }
             }
         }