+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/create/CreateNewChat.kt
@@ -15,15 +15,11 @@ import androidx.compose.material.icons.filled.GroupAdd
 import androidx.compose.material.icons.filled.TravelExplore
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -35,9 +31,9 @@ import de.connect2x.messenger.compose.view.common.Header
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.roomlist.search.SearchUsers
+import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedProgressIndicator
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.CreateNewChatViewModel
-import kotlinx.coroutines.delay
-import kotlin.time.Duration.Companion.milliseconds
 
 interface CreateNewChatView {
     @Composable
@@ -56,13 +52,6 @@ class CreateNewChatViewImpl : CreateNewChatView {
         val isCreating by createNewChatViewModel.isCreating.collectAsState()
         val error = createNewChatViewModel.error.collectAsState()
 
-        var showProgressBar by remember { mutableStateOf(false) }
-
-        LaunchedEffect(isCreating) {
-            delay(120.milliseconds)
-            showProgressBar = isCreating
-        }
-
         Box(Modifier.fillMaxSize()) {
             Box {
                 if (error.value != null) {
@@ -71,8 +60,11 @@ class CreateNewChatViewImpl : CreateNewChatView {
 
                 Column {
                     Header(createNewChatViewModel::cancel, i18n.createNewChatTitle())
-                    if (showProgressBar) {
-                        LinearProgressIndicator(Modifier.fillMaxWidth())
+                    if (isCreating) {
+                        ThemedProgressIndicator(
+                            Modifier.fillMaxWidth(),
+                            MaterialTheme.components.linearProgressIndicator
+                        )
                     }
                     AddOrSearchGroup(createNewChatViewModel)
                     HorizontalDivider(Modifier.fillMaxWidth().width(1.dp))