+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/theme/ComponentStyles.kt
@@ -7,6 +7,8 @@ import de.connect2x.messenger.compose.view.theme.components.ButtonStyle
 import de.connect2x.messenger.compose.view.theme.components.IconButtonStyle
 import de.connect2x.messenger.compose.view.theme.components.FloatingActionButtonStyle
 import de.connect2x.messenger.compose.view.theme.components.InputAreaStyle
+import de.connect2x.messenger.compose.view.theme.components.ProgressIndicatorStyle.CircularProgressIndicatorStyle
+import de.connect2x.messenger.compose.view.theme.components.ProgressIndicatorStyle.LinearProgressIndicatorStyle
 import de.connect2x.messenger.compose.view.theme.components.SurfaceStyle
 import de.connect2x.messenger.compose.view.theme.components.TooltipStyle
 
@@ -45,6 +47,11 @@ data class ComponentStyles(
     val messageReference: SurfaceStyle,
     // tooltips
     val tooltip: TooltipStyle,
+    // progress indicators
+    val circularProgressIndicator: CircularProgressIndicatorStyle,
+    val smallCircularProgressIndicator: CircularProgressIndicatorStyle,
+    val extraSmallCircularProgressIndicator: CircularProgressIndicatorStyle,
+    val linearProgressIndicator: LinearProgressIndicatorStyle,
 )
 
 internal val LocalComponentStyles = staticCompositionLocalOf<ComponentStyles> { error("compositionLocal not defined") }